+++ src/script/microbit-interfacing/MicrobitSerial.ts
